% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_name.R, R/utils.R
\name{safe_name}
\alias{safe_name}
\title{Make a safe filename}
\usage{
safe_name(x)

safe_name(x)
}
\arguments{
\item{x}{A character string to clean.}
}
\value{
A cleaned filename string.
}
\description{
Ensures file names contain only safe ASCII characters.
}
