% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSim-package.R
\docType{package}
\name{CDSim-package}
\alias{CDSim}
\alias{CDSim-package}
\title{CDSim: Climate Data Simulation Toolkit}
\description{
Tools for generating and exporting synthetic climate observation datasets.
}
\author{
Isaac Osei and Acheampong Baafi-Adomako
}
\keyword{climate}
\keyword{netcdf}
\keyword{rainfall}
\keyword{simulation}
\keyword{temperature}
